flask/sposs/requirements.txt
-i https://pypi.org/simple
-e ../../../python-utils
amqp==5.0.9; python_version >= '3.6'
billiard==3.6.4.0
celery==5.2.3
click==8.0.3; python_version >= '3.6'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask==2.0.2
flask-cors==3.0.10
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
kombu==5.2.3; python_version >= '3.7'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow==3.14.1
packaging==21.3; python_version >= '3.6'
prompt-toolkit==3.0.24; python_full_version >= '3.6.2'
pyparsing==3.0.6; python_version >= '3.6'
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
webargs==8.0.1
werkzeug==2.0.2; python_version >= '3.6'
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
